+++ modules/src/core/ics02_client/events.rs
@@ -5,7 +5,6 @@ use tendermint::abci::tag::Tag;
 use tendermint::abci::Event as AbciEvent;
 
 use crate::core::ics02_client::client_type::ClientType;
-use crate::core::ics02_client::error::Error;
 use crate::core::ics02_client::header::AnyHeader;
 use crate::core::ics02_client::height::Height;
 use crate::core::ics24_host::identifier::ClientId;
@@ -114,14 +113,6 @@ impl From<Attributes> for CreateClient {
     }
 }
 
-impl TryFrom<&AbciEvent> for CreateClient {
-    type Error = Error;
-
-    fn try_from(abci_event: &AbciEvent) -> Result<Self, Self::Error> {
-        extract_attributes_from_tx(abci_event).map(CreateClient)
-    }
-}
-
 impl From<CreateClient> for IbcEvent {
     fn from(v: CreateClient) -> Self {
         IbcEvent::CreateClient(v)
@@ -171,17 +162,6 @@ impl From<Attributes> for UpdateClient {
             common: attrs,
             header: None,
         }
-    }
-}
-
-impl TryFrom<&AbciEvent> for UpdateClient {
-    type Error = Error;
-
-    fn try_from(abci_event: &AbciEvent) -> Result<Self, Self::Error> {
-        extract_attributes_from_tx(abci_event).map(|attributes| UpdateClient {
-            common: attributes,
-            header: extract_header_from_tx(abci_event).ok(),
-        })
     }
 }
 
@@ -237,14 +217,6 @@ impl From<Attributes> for ClientMisbehaviour {
     }
 }
 
-impl TryFrom<&AbciEvent> for ClientMisbehaviour {
-    type Error = Error;
-
-    fn try_from(abci_event: &AbciEvent) -> Result<Self, Self::Error> {
-        extract_attributes_from_tx(abci_event).map(ClientMisbehaviour)
-    }
-}
-
 impl From<ClientMisbehaviour> for IbcEvent {
     fn from(v: ClientMisbehaviour) -> Self {
         IbcEvent::ClientMisbehaviour(v)
@@ -277,103 +249,12 @@ impl From<Attributes> for UpgradeClient {
     }
 }
 
-impl TryFrom<&AbciEvent> for UpgradeClient {
-    type Error = Error;
-
-    fn try_from(abci_event: &AbciEvent) -> Result<Self, Self::Error> {
-        extract_attributes_from_tx(abci_event).map(UpgradeClient)
-    }
-}
-
 impl From<UpgradeClient> for AbciEvent {
     fn from(v: UpgradeClient) -> Self {
         let attributes = Vec::<Tag>::from(v.0);
         AbciEvent {
             type_str: IbcEventType::UpgradeClient.as_str().to_string(),
             attributes,
-        }
-    }
-}
-
-fn extract_attributes_from_tx(event: &AbciEvent) -> Result<Attributes, Error> {
-    let mut attr = Attributes::default();
-
-    for tag in &event.attributes {
-        let key = tag.key.as_ref();
-        let value = tag.value.as_ref();
-        match key {
-            CLIENT_ID_ATTRIBUTE_KEY => {
-                attr.client_id = value.parse().map_err(Error::invalid_client_identifier)?
-            }
-            CLIENT_TYPE_ATTRIBUTE_KEY => {
-                attr.client_type = value
-                    .parse()
-                    .map_err(|_| Error::unknown_client_type(value.to_string()))?
-            }
-            CONSENSUS_HEIGHT_ATTRIBUTE_KEY => {
-                attr.consensus_height = value
-                    .parse()
-                    .map_err(|e| Error::invalid_string_as_height(value.to_string(), e))?
-            }
-            _ => {}
-        }
-    }
-
-    Ok(attr)
-}
-
-pub fn extract_header_from_tx(event: &AbciEvent) -> Result<AnyHeader, Error> {
-    for tag in &event.attributes {
-        let key = tag.key.as_ref();
-        let value = tag.value.as_ref();
-        if key == HEADER_ATTRIBUTE_KEY {
-            return AnyHeader::decode_from_string(value);
-        }
-    }
-    Err(Error::missing_raw_header())
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::core::ics02_client::header::Header;
-    use crate::mock::header::MockHeader;
-
-    use super::*;
-
-    #[test]
-    fn client_event_to_abci_event() {
-        let consensus_height = Height::new(1, 1).unwrap();
-        let attributes = Attributes {
-            client_id: "test_client".parse().unwrap(),
-            client_type: ClientType::Tendermint,
-            consensus_height,
-        };
-        let mut abci_events = vec![];
-        let create_client = CreateClient::from(attributes.clone());
-        abci_events.push(AbciEvent::from(create_client.clone()));
-        let client_misbehaviour = ClientMisbehaviour::from(attributes.clone());
-        abci_events.push(AbciEvent::from(client_misbehaviour.clone()));
-        let upgrade_client = UpgradeClient::from(attributes.clone());
-        abci_events.push(AbciEvent::from(upgrade_client.clone()));
-        let mut update_client = UpdateClient::from(attributes);
-        let header = MockHeader::new(consensus_height).wrap_any();
-        update_client.header = Some(header);
-        abci_events.push(AbciEvent::from(update_client.clone()));
-
-        for abci_event in abci_events {
-            match IbcEvent::try_from(&abci_event).ok() {
-                Some(ibc_event) => match ibc_event {
-                    IbcEvent::CreateClient(e) => assert_eq!(e.0, create_client.0),
-                    IbcEvent::ClientMisbehaviour(e) => assert_eq!(e.0, client_misbehaviour.0),
-                    IbcEvent::UpgradeClient(e) => assert_eq!(e.0, upgrade_client.0),
-                    IbcEvent::UpdateClient(e) => {
-                        assert_eq!(e.common, update_client.common);
-                        assert_eq!(e.header, update_client.header);
-                    }
-                    _ => panic!("unexpected event type"),
-                },
-                None => panic!("converted event was wrong"),
-            }
         }
     }
 }
