+++ modules/src/events.rs
@@ -324,73 +324,6 @@ impl TryFrom<IbcEvent> for AbciEvent {
     }
 }
 
-impl TryFrom<&AbciEvent> for IbcEvent {
-    type Error = Error;
-
-    fn try_from(abci_event: &AbciEvent) -> Result<Self, Self::Error> {
-        match abci_event.type_str.parse() {
-            Ok(IbcEventType::CreateClient) => Ok(IbcEvent::CreateClient(
-                ClientEvents::CreateClient::try_from(abci_event).map_err(Error::client)?,
-            )),
-            Ok(IbcEventType::UpdateClient) => Ok(IbcEvent::UpdateClient(
-                ClientEvents::UpdateClient::try_from(abci_event).map_err(Error::client)?,
-            )),
-            Ok(IbcEventType::UpgradeClient) => Ok(IbcEvent::UpgradeClient(
-                ClientEvents::UpgradeClient::try_from(abci_event).map_err(Error::client)?,
-            )),
-            Ok(IbcEventType::ClientMisbehaviour) => Ok(IbcEvent::ClientMisbehaviour(
-                ClientEvents::ClientMisbehaviour::try_from(abci_event).map_err(Error::client)?,
-            )),
-            Ok(IbcEventType::OpenInitConnection) => Ok(IbcEvent::OpenInitConnection(
-                ConnectionEvents::OpenInit::try_from(abci_event).map_err(Error::connection)?,
-            )),
-            Ok(IbcEventType::OpenTryConnection) => Ok(IbcEvent::OpenTryConnection(
-                ConnectionEvents::OpenTry::try_from(abci_event).map_err(Error::connection)?,
-            )),
-            Ok(IbcEventType::OpenAckConnection) => Ok(IbcEvent::OpenAckConnection(
-                ConnectionEvents::OpenAck::try_from(abci_event).map_err(Error::connection)?,
-            )),
-            Ok(IbcEventType::OpenConfirmConnection) => Ok(IbcEvent::OpenConfirmConnection(
-                ConnectionEvents::OpenConfirm::try_from(abci_event).map_err(Error::connection)?,
-            )),
-            Ok(IbcEventType::OpenInitChannel) => Ok(IbcEvent::OpenInitChannel(
-                ChannelEvents::OpenInit::try_from(abci_event).map_err(Error::channel)?,
-            )),
-            Ok(IbcEventType::OpenTryChannel) => Ok(IbcEvent::OpenTryChannel(
-                ChannelEvents::OpenTry::try_from(abci_event).map_err(Error::channel)?,
-            )),
-            Ok(IbcEventType::OpenAckChannel) => Ok(IbcEvent::OpenAckChannel(
-                ChannelEvents::OpenAck::try_from(abci_event).map_err(Error::channel)?,
-            )),
-            Ok(IbcEventType::OpenConfirmChannel) => Ok(IbcEvent::OpenConfirmChannel(
-                ChannelEvents::OpenConfirm::try_from(abci_event).map_err(Error::channel)?,
-            )),
-            Ok(IbcEventType::CloseInitChannel) => Ok(IbcEvent::CloseInitChannel(
-                ChannelEvents::CloseInit::try_from(abci_event).map_err(Error::channel)?,
-            )),
-            Ok(IbcEventType::CloseConfirmChannel) => Ok(IbcEvent::CloseConfirmChannel(
-                ChannelEvents::CloseConfirm::try_from(abci_event).map_err(Error::channel)?,
-            )),
-            Ok(IbcEventType::SendPacket) => Ok(IbcEvent::SendPacket(
-                ChannelEvents::SendPacket::try_from(abci_event).map_err(Error::channel)?,
-            )),
-            Ok(IbcEventType::WriteAck) => Ok(IbcEvent::WriteAcknowledgement(
-                ChannelEvents::WriteAcknowledgement::try_from(abci_event)
-                    .map_err(Error::channel)?,
-            )),
-            Ok(IbcEventType::AckPacket) => Ok(IbcEvent::AcknowledgePacket(
-                ChannelEvents::AcknowledgePacket::try_from(abci_event).map_err(Error::channel)?,
-            )),
-            Ok(IbcEventType::Timeout) => Ok(IbcEvent::TimeoutPacket(
-                ChannelEvents::TimeoutPacket::try_from(abci_event).map_err(Error::channel)?,
-            )),
-            _ => Err(Error::unsupported_abci_event(
-                abci_event.type_str.to_owned(),
-            )),
-        }
-    }
-}
-
 impl IbcEvent {
     pub fn to_json(&self) -> String {
         match serde_json::to_string(self) {
